lock down deleteItem by permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -42,8 +42,14 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{id, title}`);
-    //TODO check if they own it
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("you no be do this");
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   deleteManyItems: forwardTo("db"),
